updating path references after moving entries folder
